Fix: remove duplicate code
